fix(inlineerror): adding inline error on empty state
