+++ src/elements/content-sidebar/activity-feed/activity-feed/ActivityFeedItem.js
@@ -24,7 +24,6 @@ type Props = {
     approverSelectorContacts?: SelectorItems,
     created_at: string,
     created_by: User,
-
     currentUser?: User,
     getApproverWithQuery?: Function,
     getAvatarUrl: GetAvatarUrlCallback,
@@ -34,7 +33,7 @@ type Props = {
     intl: IntlShape,
     isFocused?: boolean,
     isPending: boolean,
-    item: FeedItem,
+    item: any,
     mentionSelectorContacts?: SelectorItems,
     onAppActivityDelete?: Function,
     onCommentDelete?: Function,
@@ -45,127 +44,117 @@ type Props = {
     onTaskEdit?: Function,
     onTaskModalClose?: Function,
     onVersionInfo?: Function,
+    refValue?: { current: null | HTMLElement },
     rendered_text: string,
     translations?: Translations,
     words: string,
 };
 
-const ActivityFeedItem = React.forwardRef<Props, ?HTMLElement>(
-    (
-        {
-            isPending,
-            approverSelectorContacts,
-            currentUser,
-            item,
-            mentionSelectorContacts,
-            getMentionWithQuery,
-            onAppActivityDelete,
-            onCommentDelete,
-            onCommentEdit,
-            onTaskDelete,
-            onTaskEdit,
-            onTaskAssignmentUpdate,
-            onTaskModalClose,
-            onVersionInfo,
-            translations,
-            getApproverWithQuery,
-            getAvatarUrl,
-            getUserProfileUrl,
-            isFocused,
-            activeFeedItemRef,
-        }: Props,
-        ref,
-    ): React.Node => {
-        const { type, id, permissions } = item;
-        if (!currentUser) return null;
-        switch (type) {
-            case 'comment':
-                return (
-                    <li
-                        className={classNames('bcs-activity-feed-comment', { 'bcs-is-focused': isFocused })}
-                        // { 'bcs-is-focused': isFocused }
-                        data-testid="comment"
-                        ref={activeFeedItemRef}
-                    >
-                        <Comment
-                            {...item}
-                            currentUser={currentUser}
-                            getAvatarUrl={getAvatarUrl}
-                            getMentionWithQuery={getMentionWithQuery}
-                            getUserProfileUrl={getUserProfileUrl}
-                            mentionSelectorContacts={mentionSelectorContacts}
-                            onDelete={onCommentDelete}
-                            onEdit={onCommentEdit}
-                            permissions={{
-                                can_delete: getProp(permissions, 'can_delete', false),
-                                can_edit: getProp(permissions, 'can_edit', false),
-                            }}
-                            translations={translations}
-                        />
-                    </li>
-                );
-            case 'task':
-                return (
-                    <li
-                        className={classNames('bcs-activity-feed-task-new', { 'bcs-is-focused': isFocused })}
-                        data-testid="task"
-                        ref={activeFeedItemRef}
-                    >
-                        <TaskNew
-                            {...item}
-                            approverSelectorContacts={approverSelectorContacts}
-                            currentUser={currentUser}
-                            getApproverWithQuery={getApproverWithQuery}
-                            getAvatarUrl={getAvatarUrl}
-                            getUserProfileUrl={getUserProfileUrl}
-                            onAssignmentUpdate={onTaskAssignmentUpdate}
-                            onDelete={onTaskDelete}
-                            onEdit={onTaskEdit}
-                            onModalClose={onTaskModalClose}
-                            translations={translations}
-                        />
-                    </li>
-                );
-            case 'file_version':
-                return (
-                    <li key={type + id} className="bcs-version-item" data-testid="version">
-                        {item.versions ? (
-                            <CollapsedVersion {...item} onInfo={onVersionInfo} />
-                        ) : (
-                            <Version {...item} onInfo={onVersionInfo} />
-                        )}
-                    </li>
-                );
-            case 'keywords':
-                return (
-                    <li key={type + id} className="bcs-keywords-item" data-testid="keyword">
-                        <Keywords {...item} />
-                    </li>
-                );
-            case 'app_activity':
-                return (
-                    <li key={type + id} className="bcs-activity-feed-app-activity" data-testid="app-activity">
-                        <AppActivity
-                            // activity
-                            // _template={activity_template}
-                            // app={app}
-                            // created_at={created_at}
-                            // created_by={created_by}
-                            // id={id}
-                            // intl={intl}
-                            isPending={isPending}
-                            // rendered_text={rendered_text}
-                            currentUser={currentUser}
-                            onDelete={onAppActivityDelete}
-                            {...item}
-                        />
-                    </li>
-                );
-            default:
-                return null;
-        }
-    },
-);
+const ActivityFeedItem = ({
+    isPending,
+    approverSelectorContacts,
+    currentUser,
+    item,
+    mentionSelectorContacts,
+    getMentionWithQuery,
+    onAppActivityDelete,
+    onCommentDelete,
+    onCommentEdit,
+    onTaskDelete,
+    onTaskEdit,
+    onTaskAssignmentUpdate,
+    onTaskModalClose,
+    onVersionInfo,
+    translations,
+    getApproverWithQuery,
+    getAvatarUrl,
+    getUserProfileUrl,
+    isFocused,
+    activeFeedItemRef,
+}: Props): React.Node => {
+    if (!item) {
+        return null;
+    }
+    const { type, id, permissions } = item;
+    if (!currentUser) return null;
+    switch (type) {
+        case 'comment':
+            return (
+                <li
+                    className={classNames('bcs-activity-feed-comment', { 'bcs-is-focused': isFocused })}
+                    data-testid="comment"
+                    ref={activeFeedItemRef}
+                >
+                    <Comment
+                        {...item}
+                        currentUser={currentUser}
+                        getAvatarUrl={getAvatarUrl}
+                        getMentionWithQuery={getMentionWithQuery}
+                        getUserProfileUrl={getUserProfileUrl}
+                        mentionSelectorContacts={mentionSelectorContacts}
+                        onDelete={onCommentDelete}
+                        onEdit={onCommentEdit}
+                        permissions={{
+                            can_delete: getProp(permissions, 'can_delete', false),
+                            can_edit: getProp(permissions, 'can_edit', false),
+                        }}
+                        translations={translations}
+                    />
+                </li>
+            );
+        case 'task':
+            return (
+                <li
+                    className={classNames('bcs-activity-feed-task-new', { 'bcs-is-focused': isFocused })}
+                    data-testid="task"
+                    ref={activeFeedItemRef}
+                >
+                    <TaskNew
+                        {...item}
+                        approverSelectorContacts={approverSelectorContacts}
+                        currentUser={currentUser}
+                        getApproverWithQuery={getApproverWithQuery}
+                        getAvatarUrl={getAvatarUrl}
+                        getUserProfileUrl={getUserProfileUrl}
+                        onAssignmentUpdate={onTaskAssignmentUpdate}
+                        onDelete={onTaskDelete}
+                        onEdit={onTaskEdit}
+                        onModalClose={onTaskModalClose}
+                        translations={translations}
+                    />
+                </li>
+            );
+        case 'file_version':
+            return (
+                <li key={type + id} className="bcs-version-item" data-testid="version">
+                    {item.versions ? (
+                        <CollapsedVersion {...item} onInfo={onVersionInfo} />
+                    ) : (
+                        <Version {...item} onInfo={onVersionInfo} />
+                    )}
+                </li>
+            );
+        case 'keywords':
+            return (
+                <li key={type + id} className="bcs-keywords-item" data-testid="keyword">
+                    <Keywords {...item} />
+                </li>
+            );
+        case 'app_activity':
+            return (
+                <li key={type + id} className="bcs-activity-feed-app-activity" data-testid="app-activity">
+                    <AppActivity
+                        isPending={isPending}
+                        currentUser={currentUser}
+                        onDelete={onAppActivityDelete}
+                        {...item}
+                    />
+                </li>
+            );
+        default:
+            return null;
+    }
+};
 
 export { ActivityFeedItem as ActivityFeedComponent };
 export default withErrorHandling(ActivityFeedItem);
